Mobile Anti-Theft System
Today, we live in a world of virtual reality. It makes sense to say that
none of us can live without our computers and mobiles and other gadgets.
With mobile devices becoming more and more essential to survive with,
people have started storing more and more confidential data and
information on their mobile devices. In most cases, it is the
information that happens to be more important than the computer or
mobile itself. So, what happens if your mobile or laptop gets stolen or
is lost? Wouldn't it be terrible to know that an unauthorized user has
complete access to your information? Taking all of this in mind, Mobile
Anti-Theft system is developed.
While some stolen mobile phones will wind up in pawn shops, sold on the
street corners or in online classifieds, the vast majority of stolen
devices are now being shipped internationally. International markets
(particularly countries in Africa, Latin America, and Asia pacific) pay
a premium for Smartphone’s, with the high demand pushing the prices way
up made nearly $4 million in 8 months as middleman selling phones to
Hong Kong, where they can fetch as much as $2000 per phone. (Arieanna
Schwebber).
Even in Nigeria and Ghana most Mobile users are alarmed at the rate of
GSM phone theft in the country, the experience pales into nothing when
compared with happens in South Africa. South Africa easily qualifies as
the headquarters of telephone theft on the African continent. In the
average, Vodacom and MTN receives about 10,000 phone theft a month, no
thanks to the activities of GSM phone thieves in down town Johannesburg,
Lagos and Accra who make life difficult for cell phone subscribers. [1]
MATS consist of a number of essential anti-theft features that many
other devices do not contain. It will use android built in feature to
retrieve the information about the change of SIM on your mobile phone
incase its get lost or stolen; this project will also help in tracking
down your stolen mobile with GPS and GPRS technology built in most of
the smart phones to identify the location of your mobile phone.